apps/dashboard/requirements.txt
flask>=3.0,<4
psycopg2-binary>=2.9,<3
gunicorn>=22.0,<23
